pyproject.toml
[tool.poetry]
name = "gfmodules-national-referral-index"
version = "0.1.0"
description = ""
license = "EUPL-1.2"
authors = ["Ministerie van Volksgezondheid, Welzijn en Sport"]
repository = "https://github.com/minvws/gfmodules-national-referral-index"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "0.38.0"
fastapi = "^0.120"
mypy = "^1.18.2"
pydantic-settings = "^2.11.0"
inject = "^5.3.0"
sqlalchemy = "^2.0.44"
psycopg = { extras = ["binary", "pool"], version = "^3.2.12" }
types-requests = "^2.32.4.20250809"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.24.0"
opentelemetry-instrumentation = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-instrumentation-requests = "^0.45b0"
requests = "^2.32.5"
statsd = "^4.0.1"
puzi = { git = "https://github.com/minvws/puzi-python" }
click = "^8.3.0"
pyjwt = "^2.10.1"
pyoprf = "0.8.0"
jwcrypto = "^1.5.6"
yarl = "^1.22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.0"
httpx = "^0.28.1"
ruff = "0.14.2"
codespell = "^2.4.1"
pip-audit = "^2.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
cache-dir = "~/.cache/ruff"
include = ["pyproject.toml", "app/*.py", "tests/*.py"]
line-length = 120

[tool.mypy]
files = "app,tests"
python_version = "3.11"
strict = true
cache_dir = "~/.cache/mypy"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"

# Since this are our tests, we don't need "-> None" annotations everywhere
disallow_untyped_defs = false

[tool.pytest.ini_options]
cache_dir = "~/.cache/pytest"
testpaths = ["tests"]
